backend/pyproject.toml
[project]
name = "baby-foot-elo"
version = "0.1.0"
description = "Backend pour la gestion du classement ELO de Baby Foot IHM"
authors = [
    {name = "Kone Abdallah Hakim",email = "71593784+schalappe@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "../README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "psycopg2-binary (>=2.9.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "supabase (>=2.15.2,<3.0.0)",
]

[tool.poetry]
name = "baby-foot-elo"
version = "0.1.0"
description = "Backend pour la gestion du classement ELO de Baby Foot IHM"
authors = ["Kone Abdallah Hakim <71593784+schalappe@users.noreply.github.com>"]

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
pylint = "^3.3.6"
radon = "^6.0.1"

[tool.black]
line-length = 119
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
